+++ python/ray/tune/execution/checkpoint_manager.py
@@ -3,10 +3,8 @@ import logging
 from typing import Callable, Optional
 
 from ray.tune.result import TRAINING_ITERATION
+from ray.air.config import CheckpointConfig, MIN, MAX
 from ray.util.ml_utils.checkpoint_manager import (
-    CheckpointConfig,
-    MIN,
-    MAX,
     _CheckpointManager as CommonCheckpointManager,
     _TrackedCheckpoint,
     CheckpointStorage,
